requirements.txt
keras_flops==0.1.2
matplotlib==3.5.2
numpy==1.23.1
pyserial==3.5
scikit_learn==1.2.0
